increase test coverage a bit and some clean up in general
- registered secure store and card list service into the container
- created separated files for mocks
