+++ CardFortressTests/ListViewModelTests.swift
@@ -30,8 +30,8 @@ final class ListViewModelTests: XCTestCase {
         
         //given
         var creditCards = [CreditCard]()
+
         //when
-        
         let expectation = self.expectation(description: "fetch cards")
         
         viewModel.itemsPublisher
@@ -58,8 +58,8 @@ final class ListViewModelTests: XCTestCase {
         
         //given
         let creditCard = CreditCard(number: 123, cvv: 123, date: "123", cardName: "Visa", cardHolderName: "Juan Perez")
-        //when
         
+        //when
         let expectation = self.expectation(description: "add card")
         
         viewModel.itemsPublisher
@@ -80,10 +80,9 @@ final class ListViewModelTests: XCTestCase {
     func testDeleteAllCards() {
         
         //given
-        //when
-        
         let expectation = self.expectation(description: "add card")
-        
+
+        //when
         viewModel.itemsPublisher
             .receive(on: DispatchQueue.main)
             .sink { _ in
@@ -97,36 +96,5 @@ final class ListViewModelTests: XCTestCase {
 
         waitForExpectations(timeout: 1)
 
-    }
-}
-
-class MockListService: CardListServiceProtocol {
-    var delete = false
-    var cards = [
-        CreditCard(number: 123, cvv: 123, date: "123", cardName: "Visa", cardHolderName: "Juan Perez"),
-        CreditCard(number: 1223, cvv: 1223, date: "1123", cardName: "MasterCard", cardHolderName: "Juan Perez"),
-        CreditCard(number: 1223, cvv: 1223, date: "1123", cardName: "Bank", cardHolderName: "Juan Perez")
-    ]
-    
-    func deleteAllCreditCardsFromSecureStore() -> Future<CardFortress.SecureStoreResult, Error> {
-        delete = true
-        return Future { promise in
-            promise(.success(.success))
-        }
-    }
-    
-
-    func addCreditCardToSecureStore(_ creditCard: CardFortress.CreditCard) -> Future<CardFortress.SecureStoreResult, Error> {
-        Future { [unowned self] promise in
-            cards.append(creditCard)
-            promise(.success(.success))
-        }
-    }
-
-    func getCreditCardsFromSecureStore() -> Future<[CreditCard], Error> {
-        
-        return Future { [unowned self] promise in
-            delete ? promise(.success([])) :promise(.success(cards))
-        }
     }
 }
